Improve Openshift Compatibility
Openshift constrains Docker to a greater degree than most vanilla
installations.  In particular, it does not allow root access by default
and runs containers as an effectiely random UID.  This change
pre-allocates the Maven repository the build output must be able to
write into and assigns it GID=0 and permissions bits 2575.

If the deault user is a member of group 0 (I believe this holds true,
but am not 100% certain), this will ensure that user has write access
and that all fies they create will also continue to be accessible.